catalog/requirements.txt
fastapi[all]
fire
python-dotenv
pytest
pytest-sugar
ruff
# sqlmodel
sqlalchemy
psycopg2-binary
# sentence-transformers
alembic
requests